refactor(replication): remove legacy delta compression
