+++ crates/tests/src/client_server/mod.rs
@@ -5,7 +5,6 @@ mod avian;
 mod connection;
 mod deterministic;
 mod diff;
-// mod delta;
 mod hierarchy;
 mod input;
 mod messages;
